requirements.txt
langchain 
langchain-community 
langchain-openai 
langgraph
langchainhub 
python-dotenv